Renamed test_aisdb to test_aisparser
